Simplified MemoryBlock, renamed SimpleBusError to BasicBusError, etc
