+++ emulator-hal/src/bus.rs
@@ -4,37 +4,15 @@ use crate::time::Instant;
 use core::convert::Infallible;
 use core::fmt;
 
-/// Used to translate an address from one address space into another
-pub trait FromAddress<T> {
-    /// Translate the given address into an address of type `Self`
-    fn from_address(address: T) -> Self;
-}
-
-/// Used to translate an address from one address space into another
-pub trait IntoAddress<T> {
-    /// Translate the given address into an address of type `T`
-    fn into_address(self) -> T;
-}
-
-impl<T, S> IntoAddress<T> for S
-where
-    T: FromAddress<S>,
-{
-    fn into_address(self) -> T {
-        T::from_address(self)
-    }
-}
-
 /// Represents an error that occurred during a bus transaction
 pub trait Error: fmt::Debug {}
 
-//impl<T: fmt::Debug + ?Sized> Error for T {}
+impl Error for Infallible {}
 
-// TODO should you have this?  Should you rename it?
 /// A simple pre-defined error type for bus transactions
 #[derive(Debug)]
 #[non_exhaustive]
-pub enum SimpleBusError {
+pub enum BasicBusError {
     /// A write access was requested, but the target is read-only
     ReadOnly,
 
@@ -50,10 +28,7 @@ pub enum SimpleBusError {
     Other,
 }
 
-// TODO the blanket impl covers this, but the blanket impl causes other problems
-impl Error for SimpleBusError {}
-
-impl Error for Infallible {}
+impl Error for BasicBusError {}
 
 /// Represents the order of bytes in a `BusAccess` operation
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
@@ -79,9 +54,6 @@ where
 
     /// The type of an error returned by this bus
     type Error: Error;
-
-    /// Returns the size of the addressable block that this device represents
-    //fn size(&self) -> usize;
 
     /// Read an arbitrary length of bytes from this device, at time `now`
     ///
@@ -386,79 +358,6 @@ where
     #[inline]
     fn write(&mut self, now: Self::Instant, addr: Address, data: &[u8]) -> Result<usize, T::Error> {
         T::write(self, now, addr, data)
-    }
-}
-
-/// An adapter that applies an address translation before accessing a wrapped bus object
-///
-/// This object implements the `BusAccess` trait, and takes address of type `AddressIn`,
-/// applies the provided address translation function to produce an address of type `AddressOut`,
-/// and then calls the equivalent trait method with that produced address, return the result
-pub struct BusAdapter<AddressIn, AddressOut, Bus, ErrorOut>
-where
-    AddressIn: Copy,
-    AddressOut: Copy,
-    Bus: BusAccess<AddressOut>,
-{
-    /// The underlying object implementing `BusAccess` that this object adapts
-    pub inner: Bus,
-    /// The translation function applied
-    pub translate: fn(AddressIn) -> AddressOut,
-    /// The error mapping function applied
-    pub map_err: fn(Bus::Error) -> ErrorOut,
-}
-
-impl<AddressIn, AddressOut, Bus, ErrorOut> BusAdapter<AddressIn, AddressOut, Bus, ErrorOut>
-where
-    AddressIn: Copy,
-    AddressOut: Copy,
-    Bus: BusAccess<AddressOut>,
-{
-    /// Construct a new instance of an adapter for the given `bus` object
-    pub fn new(
-        inner: Bus,
-        translate: fn(AddressIn) -> AddressOut,
-        map_err: fn(Bus::Error) -> ErrorOut,
-    ) -> Self {
-        Self {
-            inner,
-            translate,
-            map_err,
-        }
-    }
-}
-
-impl<AddressIn, AddressOut, Bus, ErrorOut> BusAccess<AddressIn>
-    for BusAdapter<AddressIn, AddressOut, Bus, ErrorOut>
-where
-    AddressIn: Copy,
-    AddressOut: Copy,
-    Bus: BusAccess<AddressOut>,
-    ErrorOut: Error,
-{
-    type Instant = Bus::Instant;
-    type Error = ErrorOut;
-
-    #[inline]
-    fn read(
-        &mut self,
-        now: Self::Instant,
-        addr: AddressIn,
-        data: &mut [u8],
-    ) -> Result<usize, Self::Error> {
-        let addr = (self.translate)(addr);
-        self.inner.read(now, addr, data).map_err(self.map_err)
-    }
-
-    #[inline]
-    fn write(
-        &mut self,
-        now: Self::Instant,
-        addr: AddressIn,
-        data: &[u8],
-    ) -> Result<usize, Self::Error> {
-        let addr = (self.translate)(addr);
-        self.inner.write(now, addr, data).map_err(self.map_err)
     }
 }
